+++ SP's_Marina_Services/SP_UPDATE_MARINA_SERVICE.sql
@@ -2,7 +2,7 @@ CREATE PROCEDURE `SP_UPDATE_MARINA_SERVICE`
 (
     _marina_service_id INT,
     _name VARCHAR(100),
-    _price FLOAT(16,8)
+    _price DECIMAL(12,4)
 )
 
 BEGIN
